pyproject.toml
[project]
name = "afad"
version = "0.1.0"
description = "Adaptive Federated Architecture Discovery"
requires-python = ">=3.10"
dependencies = [
    "fire>=0.7",
    "flwr>=1.25",
    "matplotlib>=3.10",
    "medmnist>=3.0",
    "numpy>=2.2",
    "pandas>=2.3",
    "pydantic>=2.12",
    "pyyaml>=6.0",
    "ray>=2.51",
    "rich>=13.9",
    "scikit-image>=0.25",
    "scikit-learn>=1.7",
    "torch>=2.9",
    "torchvision>=0.24",
    "tqdm>=4.67",
    "wandb>=0.24",
]

[project.optional-dependencies]
dev = [
    "poethepoet>=0.32",
    "pytest>=8.0",
    "pytest-cov>=7.0",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (formatter handles)
    "B007",  # unused loop control variable (common in ML)
    "B905",  # zip() without strict= (common in ML)
]

[tool.ruff.format]
quote-style = "double"

[tool.poe.tasks]
lint = "ruff check . && ruff format --check ."
format = "ruff check --fix . && ruff format ."
test = "pytest -v"
all = ["lint", "test"]
